+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -7,3 +7,4 @@ export 'tiktok.dart';
 export 'twitter.dart';
 export 'whatsapp.dart';
 export 'youtube.dart';
+export 'zoom.dart';