+++ test/deeplink_x_test.dart
@@ -81,6 +81,11 @@ void main() {
         expect(action, isA<App>());
       });
 
+      test('Zoom', () {
+        final action = Zoom.open();
+        expect(action, isA<App>());
+      });
+
       test('Pinterest', () {
         final action = Pinterest.open();
         expect(action, isA<App>());